+++ packages/examples/src/app/sub-modules/authentication/auth.module.ts
