+++ packages/examples/src/app/sub-modules/authentication/services/authenticator.service.ts
